zenoh_ingest/requirements.txt
psycopg2-binary
eclipse-zenoh
